Chore: 비지니스 로직 실패 처리 및 ApplyFormInputArr.ts 파일 삭제
