+++ src/app/apply/[id]/actions/uploadResumeAction.ts
@@ -1,6 +1,6 @@
 "use server";
 
-export async function uploadResumeAction(file: File) {
+export const uploadResumeAction = async (file: File) => {
   const formData = new FormData();
   formData.append("file", file);
 
@@ -12,10 +12,15 @@ export async function uploadResumeAction(file: File) {
     }
   );
 
-  if (!response.ok) {
-    throw new Error("이력서 업로드 실패");
+  if (response.status !== 201) {
+    return {
+      status: response.status,
+      message: "이력서 제출에 실패하였습니다.",
+    };
   }
 
-  const result = await response.json();
-  return result;
-}
+  return {
+    status: response.status,
+    data: await response.json(),
+  };
+};
