+++ src/app/apply/[id]/components/ApplyFormInputList.tsx
@@ -6,18 +6,75 @@ import { Path } from "react-hook-form";
 import { z } from "zod";
 import { applySchema } from "@/schema/apply/applySchema";
 import { ApplyFormInputListProps } from "@/types/apply";
+import { useState } from "react";
 
 const ApplyFormInputList = ({
-  inputArr,
   register,
   errors,
   watch,
   setValue,
-  setVisible,
   handleUploadResume,
 }: ApplyFormInputListProps) => {
   const resumeName = watch("resumeName");
   const resumeId = watch("resumeId");
+  const [visible, setVisible] = useState(false);
+
+  const inputArr = [
+    {
+      label: "이름",
+      name: "name",
+      type: "text",
+      placeholder: "이름을 입력해주세요.",
+      error: errors.name,
+      register: register("name"),
+      inputStyle: "basic",
+    },
+    {
+      label: "연락처",
+      name: "phoneNumber",
+      type: "tel",
+      placeholder: "숫자만 입력해주세요.",
+      error: errors.phoneNumber,
+      register: register("phoneNumber"),
+      inputStyle: "basic",
+    },
+    {
+      label: "경력(개월 수)",
+      name: "experienceMonths",
+      type: "number",
+      placeholder: "숫자만 입력해주세요.",
+      error: errors.experienceMonths,
+      register: register("experienceMonths"),
+      inputStyle: "basic",
+    },
+    {
+      label: "이력서",
+      name: "resumeName",
+      type: "file",
+      placeholder: "파일 업로드하기",
+      error: errors.resumeName,
+      register: register("resumeName"),
+      inputStyle: "file",
+    },
+    {
+      label: "자기 소개",
+      name: "introduction",
+      type: "text",
+      placeholder: "최대 200자까지 입력 가능합니다.",
+      error: errors.introduction,
+      register: register("introduction"),
+      inputStyle: "textarea",
+    },
+    {
+      label: "비밀번호",
+      name: "password",
+      type: visible ? "text" : "password",
+      placeholder: "비밀번호를 입력해주세요.",
+      error: errors.password,
+      register: register("password"),
+      inputStyle: "basic",
+    },
+  ];
 
   return (
     <>
@@ -89,7 +146,7 @@ const ApplyFormInputList = ({
 
           {input.name === "password" && (
             <Image
-              onClick={() => setVisible((prev: boolean) => !prev)}
+              onClick={() => setVisible(!visible)}
               src={
                 input.type === "text"
                   ? "/icon/visible.svg"
