+++ src/types/apply.ts
@@ -1,23 +1,16 @@
 import { applySchema } from "@/schema/apply/applySchema";
-import { FieldErrors, UseFormRegister } from "react-hook-form";
+import {
+  FieldErrors,
+  UseFormRegister,
+  UseFormSetValue,
+  UseFormWatch,
+} from "react-hook-form";
 import { z } from "zod";
 
-export interface ApplyFormInputData {
-  label: string;
-  name: string;
-  type: string;
-  placeholder: string;
-  inputStyle: string;
-  error: any;
-  register: any;
-}
-
 export interface ApplyFormInputListProps {
-  inputArr: ApplyFormInputData[];
   register: UseFormRegister<z.infer<typeof applySchema>>;
   errors: FieldErrors<z.infer<typeof applySchema>>;
-  watch: (fieldName: string) => any;
-  setValue: any;
-  setVisible: any;
+  watch: UseFormWatch<z.infer<typeof applySchema>>;
+  setValue: UseFormSetValue<z.infer<typeof applySchema>>;
   handleUploadResume: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
